// src/api/models/property.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * KOSMO Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { PropertyTransactionGroup } from './propertyTransactionGroup';
import type { PropertyPriceUnit } from './propertyPriceUnit';
import type { PropertyStatus } from './propertyStatus';
import type { File } from './file';

export interface Property {
  /**
   * Tên sản phẩm
   * @maxLength 255
   * @nullable
   */
  title?: string | null;
  /**
   * Nhóm giao dịch
   * @nullable
   */
  transaction_group?: PropertyTransactionGroup;
  /**
   * Vị trí
   * @maxLength 500
   * @nullable
   */
  location?: string | null;
  /**
   * Kiến trúc và tiện ích
   * @nullable
   */
  architecture?: string | null;
  price_unit?: PropertyPriceUnit;
  status?: PropertyStatus;
  /** @maxItems 20 */
  media_file_ids?: string[];
  media?: File[];
  /** Property ID */
  id?: string;
  /**
   * Mô tả bất động sản
   * @nullable
   */
  description?: string | null;
  /**
   * Thể loại bất động sản
   * @nullable
   */
  category?: string | null;
  /**
   * Hướng nhà
   * @nullable
   */
  direction?: string | null;
  /**
   * Số tầng
   * @nullable
   */
  floors?: number | null;
  /**
   * Số toilet
   * @nullable
   */
  toilets?: number | null;
  /**
   * Mặt tiền đường
   * @nullable
   */
  street_frontage?: string | null;
  /**
   * Số phòng khách
   * @nullable
   */
  living_rooms?: number | null;
  /**
   * Số phòng ngủ
   * @nullable
   */
  bedrooms?: number | null;
  /**
   * Pháp lý
   * @nullable
   */
  legal?: string | null;
  /**
   * File ID liên kết
   * @nullable
   */
  file_id?: string | null;
  /**
   * Phân tích bất động sản
   * @nullable
   */
  analysis?: string | null;
  /**
   * Giá bất động sản
   * @nullable
   */
  price?: number | null;
  /**
   * Số điện thoại nhân viên bán hàng
   * @nullable
   */
  phone_sale?: string | null;
  /**
   * Ngày tạo
   * @nullable
   */
  created_at?: string | null;
}